refactor: exception 패키지 변경, 상위 클래스 상속
