+++ backend/src/main/java/com/woowacourse/zzimkkong/exception/reservation/InvalidConferenceTimeException.java
@@ -1,0 +1,7 @@
+package com.woowacourse.zzimkkong.exception.reservation;
+
+public class InvalidConferenceTimeException extends ConflictSpaceSettingException {
+    public InvalidConferenceTimeException() {
+
+    }
+}
