+++ backend/src/main/java/com/woowacourse/zzimkkong/exception/reservation/InvalidTimeUnitException.java
@@ -1,0 +1,6 @@
+package com.woowacourse.zzimkkong.exception.reservation;
+
+public class InvalidTimeUnitException extends ConflictSpaceSettingException {
+    public InvalidTimeUnitException() {
+    }
+}
