+++ backend/src/test/java/com/woowacourse/zzimkkong/service/ManagerReservationServiceTest.java
@@ -3,8 +3,8 @@ package com.woowacourse.zzimkkong.service;
 import com.woowacourse.zzimkkong.domain.Reservation;
 import com.woowacourse.zzimkkong.domain.Space;
 import com.woowacourse.zzimkkong.dto.reservation.*;
-import com.woowacourse.zzimkkong.exception.InvalidConferenceTimeException;
-import com.woowacourse.zzimkkong.exception.InvalidTimeUnitException;
+import com.woowacourse.zzimkkong.exception.reservation.InvalidConferenceTimeException;
+import com.woowacourse.zzimkkong.exception.reservation.InvalidTimeUnitException;
 import com.woowacourse.zzimkkong.exception.authorization.NoAuthorityOnMapException;
 import com.woowacourse.zzimkkong.exception.map.NoSuchMapException;
 import com.woowacourse.zzimkkong.exception.reservation.*;
